embedded_system/CommonKnowledge/ProgrammingLanguage/Python/restful-api-basic-auth/requirements.txt
flask==2.3.3
flask-httpauth==4.8.0
python-dotenv==1.0.0
bcrypt==4.0.1